+++ format.sh
@@ -1,6 +1,6 @@
-autopep8 -i null_load.py
+autopep8 -i ./Load/null_load.py
 autopep8 -i cfs.py
-autopep8 -i data_create.py
-autopep8 -i load_sched.py
-autopep8 -i load.py
-autopep8 -i sched_data.py
+autopep8 -i ./Load/data_create.py
+autopep8 -i ./Sched_Load/load_sched.py
+autopep8 -i ./Load/load.py
+autopep8 -i ./Sched_Load/sched_data.py
