--- process_data.csv
@@ -1,2 +1,0 @@
-Process ID,Process Name,CPU times,RSS memory,VMS memory,Shared memory,Text segment,Data segment,Library code,Memory Percent,IO Connections,Open Files
-26,python3,0.08,12677120,15384576,6418432,2121728,6737920,0,0.15531332141677864,[],[]
