+++ requirements.txt
@@ -1,1 +1,7 @@
 sortedcontainers
+psutil
+csv
+pickle
+pandas
+tabulate
+typing
